feat: add `ChangeDirectory` utility class
the `ChangeDirectory` utility class provides a way to track the processes
movements around the filesystem. it also provides a way to assert
that the process is in an expected directory.
in a future  version, we will use this utility to move around the filesystem and
verify that the process is in the directory we expect before taking certain
actions